feat : 검색용 카테고리, 태그, 검색어 상태 저장용 atom 추가
