+++ src/recoil/loadingAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const isLoadingAtom = atom({
+  key: "isLoadingAtom",
+  default: false,
+});
